#1156 Сделан спринг-таймер который запускается раз в час и проверяет сколько времени осталось до наступления события mitting. Если времени осталось сутки то отправить напоминание по почте (учитывая настроики)

+++ src/main/resources/migrations/V3__init_tag.sql
@@ -1,4 +1,4 @@
-ALTER TABLE tag ADD COLUMN rating BIGINT;
+ALTER TABLE tag ADD COLUMN rating int;
 
 CREATE TABLE tag_wholikedit (
     tag_id bigint NOT NULL,
